Collapse request/response objects into the actions
