+++ server/src/main/java/org/elasticsearch/action/admin/indices/template/get/GetComponentTemplateAction.java
@@ -19,15 +19,144 @@
 
 package org.elasticsearch.action.admin.indices.template.get;
 
+import org.elasticsearch.action.ActionRequestValidationException;
+import org.elasticsearch.action.ActionResponse;
 import org.elasticsearch.action.ActionType;
+import org.elasticsearch.action.support.master.MasterNodeReadRequest;
+import org.elasticsearch.cluster.metadata.ComponentTemplate;
+import org.elasticsearch.common.Strings;
+import org.elasticsearch.common.io.stream.StreamInput;
+import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.common.xcontent.ToXContentObject;
+import org.elasticsearch.common.xcontent.XContentBuilder;
 
-public class GetComponentTemplateAction extends ActionType<GetComponentTemplateResponse> {
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+
+import static org.elasticsearch.action.ValidateActions.addValidationError;
+
+/**
+ * Action to retrieve one or more component templates
+ */
+public class GetComponentTemplateAction extends ActionType<GetComponentTemplateAction.Response> {
 
     public static final GetComponentTemplateAction INSTANCE = new GetComponentTemplateAction();
     public static final String NAME = "cluster:admin/component_template/get";
 
     private GetComponentTemplateAction() {
-        super(NAME, GetComponentTemplateResponse::new);
+        super(NAME, GetComponentTemplateAction.Response::new);
+    }
+
+    /**
+     * Request that to retrieve one or more component templates
+     */
+    public static class Request extends MasterNodeReadRequest<Request> {
+
+        private String[] names;
+
+        public Request() { }
+
+        public Request(String... names) {
+            this.names = names;
+        }
+
+        public Request(StreamInput in) throws IOException {
+            super(in);
+            names = in.readStringArray();
+        }
+
+        @Override
+        public void writeTo(StreamOutput out) throws IOException {
+            super.writeTo(out);
+            out.writeStringArray(names);
+        }
+
+        @Override
+        public ActionRequestValidationException validate() {
+            ActionRequestValidationException validationException = null;
+            if (names == null) {
+                validationException = addValidationError("names is null or empty", validationException);
+            } else {
+                for (String name : names) {
+                    if (name == null || Strings.hasText(name) == false) {
+                        validationException = addValidationError("name is missing", validationException);
+                    }
+                }
+            }
+            return validationException;
+        }
+
+        /**
+         * Sets the names of the component templates.
+         */
+        public Request names(String... names) {
+            this.names = names;
+            return this;
+        }
+
+        /**
+         * The names of the component templates.
+         */
+        public String[] names() {
+            return this.names;
+        }
+    }
+
+    public static class Response extends ActionResponse implements ToXContentObject {
+
+        private final Map<String, ComponentTemplate> componentTemplates;
+
+        public Response(StreamInput in) throws IOException {
+            super(in);
+            int size = in.readVInt();
+            componentTemplates = new HashMap<>();
+            for (int i = 0 ; i < size ; i++) {
+                componentTemplates.put(in.readString(), new ComponentTemplate(in));
+            }
+        }
+
+        public Response(Map<String, ComponentTemplate> componentTemplates) {
+            this.componentTemplates = componentTemplates;
+        }
+
+        public Map<String, ComponentTemplate> getComponentTemplates() {
+            return componentTemplates;
+        }
+
+        @Override
+        public void writeTo(StreamOutput out) throws IOException {
+            out.writeVInt(componentTemplates.size());
+            for (Map.Entry<String, ComponentTemplate> componentTemplate : componentTemplates.entrySet()) {
+                out.writeString(componentTemplate.getKey());
+                componentTemplate.getValue().writeTo(out);
+            }
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+            Response that = (Response) o;
+            return Objects.equals(componentTemplates, that.componentTemplates);
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(componentTemplates);
+        }
+
+        @Override
+        public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
+            builder.startObject();
+            for (Map.Entry<String, ComponentTemplate> componentTemplate : this.componentTemplates.entrySet()) {
+                builder.field(componentTemplate.getKey(), componentTemplate.getValue());
+            }
+            builder.endObject();
+            return builder;
+        }
+
     }
 
 }
