--- server/src/main/java/org/elasticsearch/action/admin/indices/template/get/GetComponentTemplateRequest.java
@@ -1,85 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.action.admin.indices.template.get;
-
-import org.elasticsearch.action.ActionRequestValidationException;
-import org.elasticsearch.action.support.master.MasterNodeReadRequest;
-import org.elasticsearch.common.Strings;
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
-
-import java.io.IOException;
-
-import static org.elasticsearch.action.ValidateActions.addValidationError;
-
-/**
- * Request that to retrieve one or more component templates
- */
-public class GetComponentTemplateRequest extends MasterNodeReadRequest<GetComponentTemplateRequest> {
-
-    private String[] names;
-
-    public GetComponentTemplateRequest() { }
-
-    public GetComponentTemplateRequest(String... names) {
-        this.names = names;
-    }
-
-    public GetComponentTemplateRequest(StreamInput in) throws IOException {
-        super(in);
-        names = in.readStringArray();
-    }
-
-    @Override
-    public void writeTo(StreamOutput out) throws IOException {
-        super.writeTo(out);
-        out.writeStringArray(names);
-    }
-
-    @Override
-    public ActionRequestValidationException validate() {
-        ActionRequestValidationException validationException = null;
-        if (names == null) {
-            validationException = addValidationError("names is null or empty", validationException);
-        } else {
-            for (String name : names) {
-                if (name == null || Strings.hasText(name) == false) {
-                    validationException = addValidationError("name is missing", validationException);
-                }
-            }
-        }
-        return validationException;
-    }
-
-    /**
-     * Sets the names of the component templates.
-     */
-    public GetComponentTemplateRequest names(String... names) {
-        this.names = names;
-        return this;
-    }
-
-    /**
-     * The names of the component templates.
-     */
-    public String[] names() {
-        return this.names;
-    }
-}
